refactor: use createClass to support multi-extend, avoid name conflict
